+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix log file list in admin panel [#2104](https://github.com/ViewTube/viewtube/pull/2104)
+- Use undici.request for videoplayback proxy [#2113](https://github.com/ViewTube/viewtube/pull/2113)
 
 ### Changed